修复后台心跳请求问题
