+++ summer/Info.plist
@@ -62,6 +62,7 @@
 	<string>需要使用相册</string>
 	<key>UIBackgroundModes</key>
 	<array>
+		<string>audio</string>
 		<string>remote-notification</string>
 	</array>
 	<key>UILaunchStoryboardName</key>
